move VertexBindingAndAttributes to RefreshTypes
